backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # PostgreSQL용 - 현재 SQLite 사용 중이므로 주석 처리
redis==5.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2